wip: reorganizing code to be able to extend functionality more easily
